Move batch data sources to corresponding offline stores
