+++ sdk/python/feast/repo_config.py
@@ -195,6 +195,12 @@ class FeastConfigError(Exception):
         )
 
 
+def get_data_source_class_from_type(offline_store_type: str):
+    module_name, offline_store_type = offline_store_type.rsplit(".", 1)
+    config_class_name = offline_store_type.replace("OfflineStore", "Source")
+    return get_class_from_type(module_name, config_class_name, "DataSource")
+
+
 def get_online_config_from_type(online_store_type: str):
     if online_store_type in ONLINE_STORE_CLASS_FOR_TYPE:
         online_store_type = ONLINE_STORE_CLASS_FOR_TYPE[online_store_type]
